rename xhr helper
